Add tests & support for vertical video resolutions.
Vertical video resolutions are on the rise and this library was always showing the "highest" value between width and height first. Now it keeps the order in tact and returns whatever height & width was used correctly.
Additionally, reordered the input variables to make more sense. This has the input variable order represented correctly in the output string which is always `width x height`.